Convert RoadNetwork to segments instead of nodes and edges
