+++ src/OpenSage.Game/Terrain/Roads/IRoadSegment.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections.Generic;
+using OpenSage.Graphics.Shaders;
+
+namespace OpenSage.Terrain.Roads
+{
+    internal interface IRoadSegment
+    {        
+        void GenerateMesh(
+            RoadTemplate roadTemplate,
+            HeightMap heightMap,
+            List<RoadShaderResources.RoadVertex> vertices,
+            List<ushort> indices);
+    }
+}
