+++ src/OpenSage.Game/Terrain/Roads/RoadSegmentEndPoint.cs
@@ -1,0 +1,14 @@
+﻿using System.Numerics;
+
+namespace OpenSage.Terrain.Roads
+{
+    internal class RoadSegmentEndPoint
+    {
+        public RoadSegmentEndPoint(in Vector3 position)
+        {
+            Position = position;
+        }
+
+        public Vector3 Position { get; }
+    }
+}
